extract shared flow accounting into account_flow
